More tests for GIT_DIFF_INCLUDE_UNTRACKED_CONTENT
This flag, which says that the text of untracked files should be
included in the patch output of a diff, was not tested very
thoroughly.  This added tests.

Also, passing the GIT_DIFF_INCLUDE_UNTRACKED_CONTENT flag will now
automatically turn on the GIT_DIFF_INCLUDE_UNTRACKED flag, because
why would you ask for untracked content unless you also wanted
untracked files.  As of this commit, it will not automatically
turn on GIT_DIFF_RECURSE_UNTRACKED_DIRS, instead just yielding the
content of files that would normally show in the Untracked section
of git status output.